Now game is 10x smoother!!!!! Changed potentiometer range from 0-100 to 0-1000

+++ Assets/Scripts/GameValues.cs
@@ -2,6 +2,6 @@
 
 public class GameValues : MonoBehaviour
 {
-    public int _okLevel;
-    public int _breakLevel;
+    public float _okLevel;
+    public float _breakLevel;
 }
